wip: move p2p transactions type to common
